requirements.txt
yt-dlp
openai-whisper
transformers
torch
pydub
tkinter